requirements.txt
asgiref==3.2.10
astroid==2.8.0
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
platformdirs==2.4.0
pylint==2.11.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
wrapt==1.12.1
